+++ lib/cocoapods/installer.rb
@@ -161,6 +161,7 @@ module Pod
       else
         integrate
       end
+      write_lockfiles
       perform_post_install_actions
     end
 
@@ -296,7 +297,6 @@ module Pod
       SandboxDirCleaner.new(sandbox, pod_targets, aggregate_targets).clean!
 
       update_project_cache(cache_analysis_result, target_installation_results)
-      write_lockfiles
     end
 
     def create_and_save_projects(pod_targets_to_generate, aggregate_targets_to_generate, build_configurations, project_object_version)